chore: move styles out of elements dir
